Refactor test case mocks, fix minor nitpicks
Refactored the mocks to be a part of the testcase struct and then pass
those to a `setupMocks` function which allow for custom mock behavior.
This sigificantly reduces the code for setup and allows for more
customizable mocks. Also addressing some of @levb comments on package
names for the test folders and naming convention.