Fix race condition in TestCacheStore_StoreWithOptions
The test was failing in CI due to a race condition where the cache
cleanup goroutine or lazy deletion (triggered when an expired entry
is accessed) would call Delete on the underlying store, but the test
didn't have a mock expectation for it.

Added .Maybe() mock expectation for Delete to allow it to be called
optionally without causing the test to fail.